[충돌매니저] 레이어별 충돌 완
- 열거체 비트 플래그 사용
- const _tchar* Layter 대신 _uint로 레이어를 모두 관리

+++ Client/Private/GrimReaper.cpp
@@ -71,7 +71,7 @@ _uint CGrimReaper::Tick(_double TimeDelta)
 
     // 소울 1개당 마법공격력 1% 증가
     int iMagicAttack;
-    iMagicAttack = m_iMagicAttack + m_iMagicAttack * m_iSoulNum * 0.1f;
+    iMagicAttack = int(m_iMagicAttack + m_iMagicAttack * m_iSoulNum * 0.1f);
 
     return CPlayer::Tick(TimeDelta);
 }
